refactor(models): integrate fallback models with auxiliary settings
